Javascript Improvements
- A folder was created to store the js files
- artistcards.js was created to store the cards data
- hamburguer.js was created to store the hamburguer part
- script.js file was renamed to main.js file for claruty
- main.js file was declared as 'module' type to allos import variables between files
