+++ scripts/artistcards.js
@@ -1,0 +1,40 @@
+// structure declaration to store the main attributes of each artist
+const artistcards = [
+  {
+    image: './images/billieeilish.png',
+    name: 'Billie Eilish',
+    char: 'Along with international animal protection nonprofit In Defense of Animals and Panda Voices to call for the immediate retirement of giant pandas YaYa and LeLe.',
+    quote: 'Famous Quote: "I don’t know how to function without music. When I’m not making it, I’m listening to it. It gives me courage and takes care of my mind.”',
+  },
+  {
+    image: './images/u2.png',
+    name: 'U2',
+    char: 'Their influence impacted millions of people with its charity work. Various members have done both individual charity work as well as work through the band, its efforts include poverty relief, disaster relief, health and human rights work.',
+    quote: 'Famous Quote: "Famous Quote: “Nothing is stopping you except what is inside.”',
+  },
+  {
+    image: './images/madonna.png',
+    name: 'Madonna',
+    char: 'Raising Malawi is a charity non-profit organization that was founded by Madonna and Michael Berg in 2006. It is dedicated to helping with the extreme poverty primarily through health and education programming.',
+    quote: 'Famous Quote: “A lot of people are afraid to say what they want. That is why they do not get what they want.”',
+  },
+  {
+    image: './images/coldplay.png',
+    name: 'Coldplay',
+    char: 'In the band’s early years, Coldplay was known to give 10% of their profits to charity. Coldplay teamed up with REM’s Michael Stipe for a charity single to continue aid for Hurricane Katrina victims; proceeds benefit Mercy Corps.',
+    quote: 'Famous Quote: “Nobody said it was easy. No one ever said it would be this hard.”',
+  },
+  {
+    image: './images/mariahcarey.png',
+    name: 'Mariah Carey',
+    char: 'She has donated over a $1 million to the Fresh Air Fund, which assists youth in education and career planning. She has also participated in Children Services programs to ensure that kids are raised in a stable family environment.',
+    quote: 'Famous Quote: “when you feel like hope is gone, look inside you and be strong and you will finally see the truth- that hero lies in you.”',
+  },
+  {
+    image: './images/secret.png',
+    name: 'Surprise Act',
+    char: 'Just for Bs 50 you can participate in a contest to guess which will be our surprice act. The collection will ne donated to "Aldeas SOS" organization for supporting the homeless child',
+    quote: '“I have found that among its other benefits, giving liberates the soul of the giver."',
+  },
+];
+export default artistcards;
